validações e melhorias
