+++ .byebug_history
@@ -1,4 +1,60 @@
 continue
+io.rewind
+io.read
+io.rea
+io.read
+io.rewind
+io.gets
+gets
+io
+$sdtin
+continue
+io.read
+exit
+io.gets
+io.rewind
+io.rewinf
+io.read
+exit
+$stdin.gets
+gets
+$stdin
+continue
+io.rewind
+gets
+io.gets
+io.rewind
+io.gets
+io
+continue
+gets
+io
+exit
+input
+continue
+testeee
+io.gets
+gets
+io
+continue
+testeee
+exit
+io.gets
+io
+continue
+names
+continue
+names
+continue
+names
+n
+names
+n
+exit
+name.join(',')
+name
+n
+continue
 cc
 exit
 cc
@@ -198,59 +254,3 @@ hash
 p
 continue
 row
-p
-exit
-p
-periodos
-exit
-result.first[:res].find { |a| a[:periodo] == "1930[" }
-result.first[:res]
-result.first.values
-name
-names
-result.values
-result
-response
-json_response
-decades.uniq
-decades
-exit
-decades.uniq
-decades
-exit
-result
-heading
-rows
-exit
-result
-exit
-name
-result
-result.first
-name[:res]
-name
-decade
-name
-exit
-wxit
-decade
-result
-name[:res]
-name
-exit
-result
-exit
-EXIT
-json_response.first[:res].first[:frequencia]
-json_response.first[:res].]
-json_response.first[:res].first[:periodo]
-json_response.first[:res].first
-json_response.first[:res]
-json_response.first[:nome]
-json_response.first.nome
-json_response.first
-json_response[0].nome
-json_response
-json_response.first[:res].first
-json_response.first[:res]first
-json_response.first[:res]
