+++ app.rb
@@ -1,12 +1,20 @@
 #!/usr/bin/env ruby
 
 require_relative 'config/environment'
+# require 'cli'
 
-puts "Bem vindo! Esta aplicação fornece dados da população brasileira."
+puts "BEM VINDO!!!"
+puts
+puts "Esta aplicação fornece dados sobre frequência de nomes da população brasileira."
 puts
 puts
 
+while true
+    puts
+    puts
+    puts 'Digite app --help para visualizar todos os comandos.'
+    print 'Informe qual comando deseja: '
+    input = gets.chomp
 
-
-
-CliTest.main
+    test = Cli.command_check(input)
+end
